+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v.0.0.2
 
+* Renamed RequestError to ServerError.
 * Added ability to cache API responses using an ActiveSupport::Cache::Store. Simply using URL as cache key.
 * Explicit instrumentation and logging how many elements requested from the API.